Don't double evaluate schemas in `#hawk/malli`
Fixes #12

The gist is that reader macros are passing a form to be evaluated. So if
they call `eval` we double evaluate things and that can be a _very_
subtle footgun.

In particular, malli has core functions (like `map?`) in a registry. So
when you use a function, it looks it up in that registry. Functions have
reference identity (location in memory) so if we get a different copy of
`map?` it fails to find a schema for it and it breaks.

```clojure
;; double evaluation changes things:
user=> (let [single (eval '[:vector map?])
             double (eval single)]
         {:single-identical? (identical? map? (last single))
          :double-identical? (identical? map? (last double))})
{:single-identical? true, :double-identical? false}

;; showing different locations in memory:
user=> (let [schema #hawk/malli [:map [:k [:maybe map?]]]]
         [map? (-> (.schema schema) last last last)])
[#object[clojure.core$map_QMARK___5489
         "0xe0847a9"
         "clojure.core$map_QMARK___5489@e0847a9"]
 #object[clojure.core$map_QMARK___5489
         "0x2b812add"  ;; not= to "0xe0847a9 above
         "clojure.core$map_QMARK___5489@2b812add"]]
```

And the ultimate test is that the schemas fail:

```clojure
user=> (let [schema #hawk/malli [:map [:k [:maybe map?]]]]
         (malli.core/validate (.schema schema)
                              {:k {}}))
Execution error (ExceptionInfo) at malli.core/-exception (core.cljc:138).
:malli.core/invalid-schema
user=> (let [schema [:map [:k [:maybe map?]]]]
         (malli.core/validate schema
                              {:k {}}))
true
```